Retry syscalls if they were interrupted (EINTR)
Retry each syscall according to their man pages if they
fail and the errno is set to EINTR.

This happens when a signal interrupts the syscall and it is up
to the caller to retry it again.

The exception to this is when the global variable interrupted is != 0.

We will use this global variable to mark that the user really wants to
interrupt the program (the user sent a SIGINT or a SIGTERM) so we should
not be looping against the user's will.