Skip read/write if the syscall would block
Even if we use select(2) to avoid blocking, a read/write could block
even if the fd is ready for "reading/writing" under some conditions. See
select(2).

So, we can skip the read/write if the syscall would block (for this to
work the file descriptor must have the O_NONBLOCK flag set)