Block all the signals, allow INT/TERM only in the select call
Block all the signals by default with the exception of a few like
SIGSEGV, SIGTSTP which if we block the program behaviour is undefined or
it may be harder to use by the user.

Then, set a few signals to be allowed synchronously: SIGINT, SIGTERM and
SIGPIPE.
The first two allow the user to terminate tiburoncin; the last signals a
broken pipe but it is currently ignored.

The "synchronous handling" of the signals is done by pselect (a select
syscall without race conditions).

Currently, the first part of tiburoncin (the connection to and the
accept of a connection from) doesn't support pselect and the SIGINT and
SIGTERM are still blocked by that time which may prevent the user to
cancel the execution during that phase.