Support signal mask setting on connect and accept atomically
Set the sockets (file descriptors) non blocking
(set_nonblocking is private now).
Then, wrap the syscalls connect and accept with function (pconnect and
paccept) that trigger the syscall without blocking and waits for its
completion with the pselect syscall that supports setting a signal
mask atomically.

Tiburoncin now block all the signals but unblock SIGINT and SIGTERM in
the only blocking syscall in the code: pselect.
This allows the reception of those signals in a synchronous way.

See the details in the pconnect and paccept functions, and in the
signal.h module.