Fix unintended char to int promotion in the printf
When we print (to stdout or to file) each byte we used to do '%02x'
as a format string.

But because the buffer is a buffer of signed chars, when a negative
number is printed (like -1 or 0xff), printf expands this to an integer
and the resulting printed number is 0xffffffff instead of 0xff.

Use '%02hhx' fixes this. (doing a printf('%02x', byte & 0xff) also does
the trick).

Closes #1